backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
Pillow
gspread
python-dateutil
django-cors-headers
redis
django-redis
django-filter
